Refactor functions for doubles

+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToBooleanFunction.java
@@ -3,16 +3,15 @@ package ar.com.kfgodel.function.intertype.doubles;
 import java.util.function.Function;
 
 /**
- * This type defines an operator for primitive boolean types
  * Created by kfgodel on 23/07/17.
  */
-public interface DoubleToBooleanFunction extends Function<Double, Boolean> {
+public interface DoubleToBooleanFunction extends Function<Integer, Boolean> {
 
   boolean apply(double aBoolean);
 
   @Override
-  default Boolean apply(Double aBoolean){
-    return this.apply(aBoolean.doubleValue());
+  default Boolean apply(Integer aBoolean){
+    return this.apply(aBoolean.intValue());
   }
 
 }
