+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToByteFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.doubles;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a byte as result
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToByteFunction extends Function<Double, Byte> {
+public interface DoubleToByteFunction extends Function<Integer, Byte> {
 
   byte apply(double input);
 
   @Override
-  default Byte apply(Double input){
-    return apply(input.doubleValue());
+  default Byte apply(Integer input){
+    return apply(input.intValue());
   }
 }
