+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToIntFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.doubles;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a byte as result
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToIntFunction extends Function<Double, Integer> {
+public interface DoubleToIntFunction extends Function<Integer, Integer> {
 
   int apply(double input);
 
   @Override
-  default Integer apply(Double input){
-    return apply(input.doubleValue());
+  default Integer apply(Integer input){
+    return apply(input.intValue());
   }
 }
